Forward-ported [13429] (I'm not sure why I was coding in adium-0.8 in the first place, actually):  * Fixed up some float/int desparities which were likely causing #672. Fixed #672.  * Speak Text action can now include %n (user name), %a (alias), %m (message), and %t (time), allowing the user much more control over what gets spoken and when. Closes #142  * Sounds from sound sets are now in separate submenus in the Play Sound action, making the menu much more usable particularly when many .AdiumSoundsets are installed. Fixes #762

+++ Source/AISoundController.m
@@ -52,7 +52,7 @@
 - (NSArray *)voices{
 	return [adiumSpeech voices];
 }
-- (void)speakDemoTextForVoice:(NSString *)voiceString withPitch:(float)pitch andRate:(int)rate{
+- (void)speakDemoTextForVoice:(NSString *)voiceString withPitch:(float)pitch andRate:(float)rate{
 	[adiumSpeech speakDemoTextForVoice:voiceString withPitch:pitch andRate:rate];
 }
 - (float)defaultRate{
